requirements.txt
asgiref==3.8.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==5.0.3
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.1
django-rest-swagger==2.2.0
djangorestframework==3.15.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
executing==2.0.1
idna==3.6
inflection==0.5.1
ipython==8.22.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
openapi-codec==1.3.2
packaging==24.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
